Clean up the maze of #ifdefs in the OS code a bit.
All OS functions are always defined now, unimplemented functions will
just be stubs.

Replace OS-specific headers with os-private.h which won't pull in all
sorts of nasty system-specific headers. Kill the namespace pollution
define while we're at it.

The OS-specific API is changed so that opendisk and getparams
functions now return INT_MAX if the function is a stub. This is
because unlike listdev functions, a negative return value means errno
has always been set to something suitable for strerror().